Improve organization
