--- force-app/main/default/lwc/objectSelector/objectSelector.html
@@ -1,3 +1,0 @@
-<template>
-    <!-- This component has no visible UI - it registers a custom property editor with SurveyJS -->
-</template>
