--- force-app/main/default/lwc/objectSelector/objectSelector.js-meta.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>62.0</apiVersion>
-    <isExposed>false</isExposed>
-    <description>Utility component that registers an Object Selector property editor with SurveyJS</description>
-</LightningComponentBundle>
