+++ force-app/main/default/lwc/surveyProperty/surveyProperty.js-meta.xml
