+++ force-app/main/survey-properties/lwc/objectSelector/objectSelector.html
