+++ force-app/main/survey-properties/lwc/uuidDisplay/uuidDisplay.html
@@ -1,0 +1,3 @@
+<template>
+    <!-- This component has no visible UI - it registers a custom property editor with SurveyJS -->
+</template>
